feat(util.isDraggable): returns if element can be dragged
SOLUTION IS TEMPORARY

This method returns if element can be dragged on the page. This method
is to prevent bug when multitouch elements lose focus and user start
dragging <html> or <body>.
Method checks presence of class 'ball', for bigger portability we should
check instead that this object is among MTObjects array.